\name{AffyGeneSTPDInfo-class}
\docType{class}
\alias{AffyGeneSTPDInfo-class}

\title{Class AffyGeneSTPDInfo}
\description{Short description}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("AffyGeneSTPDInfo", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{getdb}:}{Object of class \code{"function"} }
    \item{\code{tableInfo}:}{Object of class \code{"data.frame"} }
    \item{\code{geometry}:}{Object of class \code{"integer"} }
    \item{\code{manufacturer}:}{Object of class \code{"character"} }
    \item{\code{genomebuild}:}{Object of class \code{"character"} }
  }
}
\section{Extends}{
  Class \code{"\linkS4class{AffyExpressionPDInfo}"}, directly.
  Class \code{"\linkS4class{ExpressionPDInfo}"}, by class "AffyExpressionPDInfo", distance 2.
  Class \code{"\linkS4class{DBPDInfo}"}, by class "AffyExpressionPDInfo", distance 3.
  Class \code{"\linkS4class{PDInfo}"}, by class "AffyExpressionPDInfo", distance 4.
}
\section{Methods}{
No methods defined with class "AffyGeneSTPDInfo" in the signature.
}
\author{RS}
\seealso{
}
\examples{
showClass("AffyGeneSTPDInfo")
}
\keyword{classes}
