\name{chromosome}
\alias{chromosome}
\alias{chromosome,SnpLevelSet-method}
\title{Accessor for chromosome information}
\description{
  Returns chromosome information.
}
\usage{
chromosome(object)
}
\arguments{
  \item{object}{object inheriting from \code{SnpLevelSet}}
}
\value{
  Vector with chromosome information.  Chromosomes are recorded as
  character strings (e.g., "1", "2", "X", ...)
}
\author{R. Scharpf}
\keyword{manip}
