\name{createFF}
\alias{createFF}

\title{
  Create ff objects.
}
\description{
  Creates ff objects (array-like) using settings (path) defined by oligoClasses.
}
\usage{
createFF(name, dim, vmode = "double")
}

\arguments{
  \item{name}{Prefix for filename.}
  \item{dim}{Dimensions.}
  \item{vmode}{Mode.}
}

\value{
  ff object.
}

\note{
  This function is meant to be used by developers.
}

\seealso{
  ff
}

\keyword{manip}
