\name{genomeBuild}
\alias{genomeBuild}
\alias{genomeBuild,PDInfo-method}
\alias{genomeBuild,FeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Genome Build Information}
\description{
  Returns the genome build information
}
\usage{
genomeBuild(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{PDInfo} object.}
}
\keyword{manip}

