\name{mmindex}
\alias{mmindex}
\alias{mmindex,FeatureSet-method}
\alias{mmindex,DBPDInfo-method}
\alias{pmindex}
\alias{pmindex,FeatureSet-method}
\alias{pmindex,DBPDInfo-method}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract PM or MM indexes}
\description{
  Extracts the indexes for PM or MM probes, which can be used with the
  intensity matrix for subsetting.
}
\usage{
mmindex(object, ...)
pmindex(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{AffySNPPDInfo}, \code{FeatureSet} object}
  \item{...}{extra arguments}
}
\keyword{manip}
