\name{manufacturer-methods}
\docType{methods}
\alias{manufacturer}
\alias{manufacturer-methods}
\alias{manufacturer,PDInfo-method}
\alias{manufacturer,FeatureSet-method}
\title{Manufacturer ID for FeatureSet-like objects.}
\description{
  Manufacturer ID for FeatureSet-like objects.
}
\section{Methods}{
\describe{
\item{object = "FeatureSet"}{Manufacturer ID}
\item{object = "PDInfo"}{Manufacturer ID}
}}
\keyword{methods}

