\name{oligoSnpSet-methods}
\docType{methods}
\alias{oligoSnpSet-class}
\alias{initialize,oligoSnpSet-method}
\alias{copyNumber,oligoSnpSet-method}
\alias{copyNumber<-,oligoSnpSet,matrix-method}
\alias{calls,oligoSnpSet-method}
\alias{calls<-,oligoSnpSet,matrix-method}
\alias{cnConfidence,oligoSnpSet-method}
\alias{cnConfidence<-,oligoSnpSet,matrix-method}
\alias{callsConfidence,oligoSnpSet-method}
\alias{callsConfidence<-,oligoSnpSet,matrix-method}
\title{Methods for oligoSnpSet class}
\description{Methods for oligoSnpSet}
\keyword{methods}
