\name{PDInfo-class}
\docType{class}
\alias{PDInfo-class}
\alias{AffyExpressionPDInfo-class}
\alias{AffyTilingPDInfo-class}
\alias{AffyExonPDInfo-class}
\alias{AffySTPDInfo-class}
\alias{SNPPDInfo-class}
\alias{SNPCNVPDInfo-class}
\alias{NgsTilingPDInfo-class}
\alias{TilingPDInfo-class}
\alias{ExpressionPDInfo-class}
\alias{NgsExpressionPDInfo-class}

\title{Classes "PDInfo"}
\description{Containers for chip information. Objects of this class are
  created with pdInfoBuilder, but for the moment, only a few users are
  expected to create such objects.}
\section{Objects from the Class}{
}
\section{Slots}{
	 \describe{
    \item{\code{manufacturer}:}{Object of class \code{"character"}  }
    \item{\code{genomebuild}:}{Object of class \code{"character"}  }
    \item{\code{geometry}:}{Object of class \code{"integer"} with length
    2 (rows x columns)}
  }
}
\section{Methods}{
No methods defined with class "PDInfo" in the signature.
}
\keyword{classes}
