\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,FeatureSet-method}

\title{Sample names for FeatureSet-like objects}
\description{
  Returns sample names for FeatureSet-like objects.
}
\usage{
sampleNames(object)
}

\arguments{
  \item{object}{\code{FeatureSet} object.}
}

\keyword{manip}

