\name{mmSequence}
\alias{mmSequence}
\alias{mSequence,DBPDInfo-method}
\alias{mmSequence,FeatureSet-method}
\alias{pmSequence}
\alias{pmSequence,DBPDInfo-method}
\alias{pmSequence,FeatureSet-method}

\title{Probe Sequeces}
\description{
  Accessor to the probe sequences.
}
\usage{
mmSequence(object)
pmSequence(object, ...)
}
\arguments{
  \item{object}{\code{FeatureSet} or \code{DBPDInfo} object}
  \item{...}{additional arguments}
}
\value{
  A character vector with the PM or MM probe sequences.
}

\keyword{manip}
