\name{show}
\alias{show}
\alias{show,DBPDInfo-method}
\alias{show,FeatureSet-method}
\title{Show object basic properties}
\description{
  Show FeatureSet-like and DBPDInfo-like objects basic properties.
}
\usage{
show(object)
}
\arguments{
  \item{object}{\code{FeatureSet} or \code{PDInfo} object.}
}
\keyword{manip}
