\name{SnpCallSet-methods}
\docType{methods}
\alias{SnpCallSet-class}
\alias{calls,SnpCallSet-method}
\alias{calls<-,SnpCallSet,matrix-method}
\alias{callsConfidence,SnpCallSet-method}
\alias{callsConfidence<-,SnpCallSet,matrix-method}
\title{SnpCallSet methods}
\description{Methods for SnpCallSet}
\keyword{methods}
