\name{SnpCopyNumberSet-methods}
\docType{methods}
\alias{SnpCopyNumberSet-class}
\alias{copyNumber,SnpCopyNumberSet-method}
\alias{copyNumber<-,SnpCopyNumberSet,matrix-method}
\alias{cnConfidence,SnpCopyNumberSet-method}
\alias{cnConfidence<-,SnpCopyNumberSet,matrix-method}
\title{Methods for SnpCopyNumberSet}
\description{Methods for SnpCopyNumberSet}
\keyword{methods}
