\name{SnpLevelSet-class}
\docType{class}
\alias{SnpLevelSet-class}
\alias{copyNumber,SnpLevelSet-method}
\alias{copyNumber<-,SnpLevelSet,matrix-method}
\alias{cnConfidence,SnpLevelSet-method}
\alias{cnConfidence<-,SnpLevelSet,matrix-method}
\alias{calls,SnpLevelSet-method}
\alias{calls<-,SnpLevelSet,matrix-method}
\alias{callsConfidence,SnpLevelSet-method}
\alias{callsConfidence<-,SnpLevelSet,matrix-method}
\alias{chromosome<-}
\alias{chromosome<-,SnpLevelSet,character-method}
\alias{chromosome<-,AnnotatedDataFrame,character-method}
\alias{combine,SnpLevelSet,SnpLevelSet-method}
\title{SnpLevelSet class}
\description{Virtual class for SNP-level data}
\keyword{classes}
