\name{CopyNumberSet-methods}
\docType{methods}
\alias{CopyNumberSet-methods}
\alias{copyNumber}
\alias{copyNumber,CopyNumberSet-method}
\alias{copyNumber<-,CopyNumberSet,matrix-method}
\alias{cnConfidence,CopyNumberSet-method}
\alias{cnConfidence<-,CopyNumberSet,matrix-method}
\alias{initialize,CopyNumberSet-method}
\title{  Methods for class CopyNumberSet.}
\description{
	Accessors and CopyNumberSet
}

\usage{

copyNumber(object)
cnConfidence(object)

}

\arguments{

\item{object}{\code{CopyNumberSet} object}

}

\value{

	\code{copyNumber} returns a matrix of copy number estimates.
	
	\code{cnConfidence} returns a matrix of confidence scores for
	the copy number estimates.

}
\keyword{methods}
\keyword{manip}

