\name{eSet-methods}
\docType{methods}
\alias{chromosome}
\alias{chromosome,eSet-method}
\alias{chromosome}
\alias{chromosome<-}
\alias{chromosome<-,eSet-method}
\alias{position,eSet-method}
\alias{isSnp}
\alias{isSnp,eSet-method}
\title{Accessors for eSet extensions}
\description{

	Accessors for variables stored in the featureData slot of a
	class inheriting from eSet. 
	
}
\section{Methods}{
\describe{
\item{\code{signature(object = "eSet")}}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{manip}
