\name{genomeBuild}
\alias{genomeBuild}
\alias{genomeBuild,DBPDInfo-method}
\alias{genomeBuild,FeatureSet-method}

\title{Genome Build Information}
\description{
  Returns the genome build information. This information comes from the
  annotation package and is given as an argument during the package
  creation process.
}
\usage{
genomeBuild(object)
}

\arguments{
  \item{object}{\code{PDInfo} or \code{FeatureSet} object.}
}

\keyword{manip}

